+++ app/helpers/pageflow/admin/users_helper.rb
@@ -1,0 +1,9 @@
+module Pageflow
+  module Admin
+    module UsersHelper
+      def collection_for_user_roles
+        User.roles_accessible_by(current_ability).index_by { |role| t(role, :scope => 'admin.users.roles') }
+      end
+    end
+  end
+end