+++ app/helpers/pageflow/application_helper.rb
@@ -1,4 +1,0 @@
-module Pageflow
-  module ApplicationHelper
-  end
-end