+++ lib/pageflow/user_mixin.rb
@@ -37,6 +37,10 @@ module Pageflow
       [first_name, last_name] * " "
     end
 
+    def formal_name
+      [last_name, first_name] * ", "
+    end
+
     def update_with_password(attributes)
       if needs_password?(attributes)
         super(attributes)